# Conflicts:
#	Ragon/Sources/Game/GameRoom.cs

+++ Ragon/Sources/Server/ISocketServer.cs
@@ -5,7 +5,7 @@ public interface ISocketServer
   public void Start(ushort port, int connections, uint protocol);
   public void Process();
   public void Stop();
-  public void Send(uint peerId, byte[] data, DeliveryType type);
-  public void Broadcast(uint[] peersIds, byte[] data, DeliveryType type);
-  public void Disconnect(uint peerId, uint errorCode);
+  public void Send(ushort peerId, byte[] data, DeliveryType type);
+  public void Broadcast(ushort[] peersIds, byte[] data, DeliveryType type);
+  public void Disconnect(ushort peerId, uint errorCode);
 }
